Made all test uniform using Specs2 instead of ScalaTest.
